[클라이언트] 애니메이션 구현
- 컴포넌트 패턴으로 작성
  - 베이스 컴포넌트 같은건 없고, 정말 코드 나누기 용도로 사용함

+++ Client/GameObject.h
@@ -10,7 +10,7 @@ private:
 	struct cbGameObject
 	{
 		DirectX::XMFLOAT4X4 worldMatrix;
-		FLOAT layer;
+		INT layer;
 		FLOAT alpha;
 		BOOL isFliped;
 		FLOAT dummy;
